IIMApiClient class

Provides the end-to-end queries required to update all details about the live acoustic telemetry system in place

Methods

Public methods

IIMApiClient$new()

IIMApiClient$post_data_to_api()

IIMApiClient$delete_data_from_api()

IIMApiClient$get_from_api()

IIMApiClient$get_from_api_as_dataframe()

IIMApiClient$clone()

Method new()

Usage

IIMApiClient$new(api_url)

Method post_data_to_api()

Post a nested list to the database through an API.

Usage

IIMApiClient$post_data_to_api(obj, end_point)

Arguments

obj

a nested list resulting from applying NestedListBuilder

end_point

string with the name of the end point being queried. Default "tags_with_sensors/"

Method delete_data_from_api()

Delete a single record from the database through the API

Usage

IIMApiClient$delete_data_from_api(obj, end_point)

Arguments

obj

a flat list with the ids or any unique field that identifies the record to be erased

end_point

string with the name of the end point being queried. Default "tags_with_sensors/"

Method get_from_api()

Make a request to the API and return the data in a json object.

Usage

IIMApiClient$get_from_api(end_point)

Arguments

end_point

string with the name of the end point being queried. Default "tags_with_sensors/"

obj

a nested list resulting from applying NestedListBuilder

Method get_from_api_as_dataframe()

Usage

IIMApiClient$get_from_api_as_dataframe(end_point, parameters = NULL)

Method clone()

The objects of this class are cloneable with this method.

Usage

IIMApiClient$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.