TemploApi class

API client to consume TEMPLO's API Add a tag with a sensor details to the database.

Value

TRUE or FALSE

TRUE or FALSE

record with the data being sent plus the unique identifier generated by the remote database

a dataframe with the sought information.

a dataframe with the sought information.

a dataframe with the sought information.

a dataframe with the sought information.

TRUE if successfully completed. Error message otherwise.

Methods

Public methods

TemploApiClient$new()

TemploApiClient$add_tag_with_sensor()

TemploApiClient$add_species()

TemploApiClient$add_tagged_individuals()

TemploApiClient$add_single_record_and_fetch_id()

TemploApiClient$get_sensor_types()

TemploApiClient$get_tagged_individuals()

TemploApiClient$get_environmental_readings()

TemploApiClient$get_dataframe_end_point()

TemploApiClient$delete_single_record()

TemploApiClient$clone()

Method new()

Usage

TemploApiClient$new(api_url)

Method add_tag_with_sensor()

Usage

TemploApiClient$add_tag_with_sensor(
  df,
  tag_prefix,
  sensor_prefix,
  field_names,
  end_point = "tags_with_sensors/"
)

Arguments

end_point

string with the name of the end point being queried. Default "tags_with_sensors/"

obj

a nested list resulting from applying NestedListBuilder

Method add_species()

Usage

TemploApiClient$add_species(df, end_point = "species/")

Method add_tagged_individuals()

Usage

TemploApiClient$add_tagged_individuals(
  df,
  individual_prefix,
  tag_prefix,
  sensor_prefix,
  tagged_individual_prefix,
  field_names = c("individual_data", "tag_data", "sensors_data",
    "tagged_individual_data"),
  end_point = "tagged_individuals/"
)

Arguments

df

dataframe with pontetial multiple-row individuals (if multiple sensors); one row otherwise.

individual_prefix

string identifying what columns belong to the individual dataset

tag_prefix

string identifying what columns belong to the tag dataset

sensor_prefix

string identifying what columns belong to the sensor dataset

tagged_individual_prefix

string identifying what columns belong to the tagged_individual dataset

Method add_single_record_and_fetch_id()

General method to add a single record to the database through the API

Usage

TemploApiClient$add_single_record_and_fetch_id(end_point, data)

Arguments

end_point

string with the name of the end point being queried.

data

a flat list with the fields of the record being sent.

Method get_sensor_types()

Get all sensor types currently in the database

Usage

TemploApiClient$get_sensor_types(end_point = "sensortypes/")

Arguments

end_point

string with the name of the end point being queried.

Method get_tagged_individuals()

Get all tagged individuals from the database. This operation may take a few seconds.

Usage

TemploApiClient$get_tagged_individuals(
  end_point,
  tagged_individual_fields = TAGGED_INDVIDUAL_FIELDS
)

Arguments

end_point

string with the name of the end point being queried.

tagged_individual_fields

list of sections and fields per section to be retrieved.

Method get_environmental_readings()

Get environmental data from the API that were emitted by receptors

Usage

TemploApiClient$get_environmental_readings(
  end_point = "environmental_readings/data_exports/",
  from_date,
  to_date
)

Arguments

end_point

string with the name of the end point being queried.

from_date

a yms format string start date

to_date

a yms format string end_date

Method get_dataframe_end_point()

General method to get data from the API in a dataframe format.

Usage

TemploApiClient$get_dataframe_end_point(end_point, parameters = NULL)

Arguments

end_point

string with the name of the end point being queried.

parameters

a named list with potential parameters to pass on to the API query. Default is NULL

Method delete_single_record()

General method to delete data from the API.

Usage

TemploApiClient$delete_single_record(end_point, data)

Arguments

end_point

string with the name of the end point being queried.

data

a flat list with the unique fields identifying the record being deleted

Method clone()

The objects of this class are cloneable with this method.

Usage

TemploApiClient$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.